Update 04-models-with-subtypes.json
Showing examples of using both json-schema validation and the an 'extends' keyword extension to define swagger model inheritance. In theory all the pets should get a UI display with a merge of their properties and the name property from Pet. I suggest we expand on the previous behaviour of swagger to also include attributes from the allOf in visualizations to avoid unnecessary repetition purely for swagger UI.