Add '$schema' property at root of the spec 2.x
JSON schema specifies a standard way for a document to tell what JSON
Schema document can be used to validate it. This is the '/$schema'
property.

Unfortunately this property is strictly disallowed in the Swagger 2.0
JSON schema ('/additionalProperties' is set to false): Swagger authors
decided to use instead the '/swagger' property. This is enough for
humans, but the presence of '/$schema' could be useful for machines to
validate Swagger specifications that use vendor extensions and for which
a JSON schema specifying the schema of those extensions is available.

So we add '/$schema', with a default value of
http://swagger.io/schema/v2.0/schema.json (see
[swagger-api/swagger.io#48](https://github.com/swagger-api/swagger.io/issues/#48))

As this feature is incompatible with the 2.0 schema
('/$schema' is forbidden in 2.0 because '/additionalProperties' is set
to false), setting this value in a document will only be possible if a
version 2.1 of the specification is published with that extension.