schema 2.0: fix definitions of 'consumes'/'produces'
A JSON reference ($ref) is a JSON object containing ONLY the `$ref`
property. The [specification (still a draft) for JSON
references](https://tools.ietf.org/html/draft-pbryan-zyp-json-ref-03#section-3)
says explicitely that any other property must be ignored:

> Any members other than "$ref" in a JSON Reference object SHALL be ignored.

In the schema for Swagger 2.0 the properties `consumes` and `produces`
were not respecting this contraint. This is fixed by using an `allOf`
schema that combines the `{ "$ref": ...}` and an other object containing
the other properties.