The "Common Name" field was confusing people
So many people were considering dateTime, etc to be an OpenAPI type that tool vendors were implementing functionality to normalize this.

https://github.com/mikunn/openapi-schema-to-json-schema/issues/4

I too in the past have thought I was meant to use dateTime in the format, not date-time. Let's just get rid of this column, it serves no real purpose as far as I can tell.