edit OpenAPI Specification license info
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit creates a LICENSE file that stores the full text of the
Apache license, referencing The Linux Foundation as the copyright
holder. It also updates package.json with the correct SPDX license ID
and updates the README "License" header to "Licensing".

Collectively, these changes allow Licensee to successfully identify the
license type of the OpenAPI Specification project as Apache.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>