Discriminator 3.2: propertyName should not be required in object
The `discriminator` keyword is an annotation. It doesn't affect
validation, so I would not be ideal for it to cause a failure if not
used properly. This changes the behavior of `discriminator` from being
an error if the propertyName is not a required property in the object to
being a no-op. This is also more consistent with other JSON Schema
keywords that ignore things that don't apply to them, like `properties`
only applying to instances that are objects.