Discriminator 3.2: Don't allow schema names in mapping
There is no way for implementations to tell the difference between a
value that is intended to be a schema name and a value that is expected
to be a URI. Requiring all mappings to use URIs avoids this problem
while not losing any functionality. The only consequence is that some
mappings would be a little more verbose than they otherwise would have
needed to be (example: `#/components/schemas/Dog` vs `Dog`).