Add links for this repository, turn off "blank"
This adds links for things we actually want people to do in this
repository.  It directs enhancement requests to discussions rather
than issues, mirroring what we're doing for the sig-moonwalk repo.

This is done in the config file rather than with issue templates
because GitHub forces issue templates to display at the top of the
list, and the whole point of this configuration is to route people
to other repositories for certain things.