Clarify resolving implicit connections (3.1.1)
This clarifies how to handle resolving implicit (non-URI-based)
connections in multi-document OpenAPI Descriptions.  While the
behavior is implementation-defined overall, two approaches are
RECOMMENDED for tool developers and OAD authors: one that is
best suited for sharing components among OADs, and one that is
compatible with pre-3.1 parsing approaches and is convenient
for breaking a monolithic document up purely for author convenience.

Note that the term "complete OpenAPI document" has been defined
in another change pending approval on the 3.0.4 branch.