Clarify constraints on Param Obj fields (3.1.1)
This makes the relational requirements of `content` vs `schema`
and associated fields more clear.  The mutual exclusivity requirement
is moved up to the top of the Fixed Fields section, and each table
table is given a subsection clarifying its meaning.  This makes it
easier to follow the multi-table format, which is unusual within the OAS.

The commentary on `example` and `examples` is moved to the section
on `schema` and related parameters as it appears to only apply
in that case.