Clarify allowEmptyValue further
Make it clear that the default value does not do anything, and the
non-default value (true) changes the normal expected behavior of
the "schema" or "content" fields.  This also explains why the
keyword is safe to deprecate, as "schema" and "content" can be
used to allow empty values in a more typical way.