Correctly handle appendicies in md2html
We have an Appendix A that was getting treated as a numbered section,
with a second Appendix A generated for references.

This now correctly detects sections with "Appendix" in the name,
and assigns letters appropriately.  The generated references
section is now Appendix B (and will always be the last appendix).