Encoding Obj contentType takes a media range (3.0.4)
The "the two types" appears to have come from an example involving
two types, and the TSC agrees that there was no intention to limit
it to two.  The syntax described also matches media type ranges,
which is what is used for keys under `content` keywords that take
objects of Media Type Objects.  As that is analogous to the use
of `contentType` in the Encoding Object, this change aligns the wording.

The guidance on default values was difficult to format in a readable
way inside of the fixed fields table, so this change also moves it
to its own table.