Further clarify discriminator usage (3.0.4)
This backports changes made to the Discriminator Object
in 3.1.1 and merges them with the changes  made on this branch.

Note that the previouc changes incorporated reworked ideas from
@jdesrosiers regarding the ambiguous `mapping` syntax submitted
in the 3.1.1 changes but not accepted at that time due to
compatibility concerns.  This commit merges in the parts of
that 3.1.1 change that were accepted.

For the same compatibility reasons, the MUST wording for
requiring the named discriminator property in the schema
was (regrettably) weakened to a "SHOULD but otherwise undefined",
as we have done for other problematic ambiguities.

Co-authored-by: Jason Desrosiers <jdesrosi@gmail.com>