Indirect->Implicit, unify operationId, more tweaks
Every time I try to reference this section in another PR I type
"implicit" instead of "indirect" so let's just go with that.

Let's also treat `operationId` uniformly, because it fits better
with how we parse documents, and the only thing this changes
is in the only-the-entry-document-has-an-OpenAPI-object case,
where it incorporates Path Item Objects in the entry document
that are otherwise un-referenced.  It seems likely that one would
only include Path Item Objects in the entry document's components
if one expected them to be used.