Clarify confusing use of YAML "JSON Schema"
When we mention YAML's "Failsafe schema" we give it a lower-case
"schem", as the YAML documentatio does.  We also prefix it with
"YAML".

However, we capitalize "Schema" in "JSON Schema ruleset",
which (given how much JSON Schema is used in the OAS) is a jarring
overlap with "JSON Schema".

This change aligns "YAML JSON schema ruleset" with
"YAML Failsafe ruleset" and explicitly calls out that it is
unrelated to JSON Schema.