Further clarify style+explode examples (3.2.0 port of 3859 2/3)
This aligns all examples with RFC6570 operator prefixing behavior,
which was previously only shown for `matrix` and `label`.  The
non-RFC6570 styles (`spaceDelimited`, `pipeDelimited`, and
`deepObject`) are treated as analogues of `form` and therefore
prefixed with a `?`.  The lack of suitablity of this for
cookie parameters has been addressed with an appendix in
another change, and the appendix has been stubbed out here to
ensure that the link is valid.

Switch the "empty" column heading to "undefined" to align with
RFC6570 and make clear that it is not about `allowEmptyValue`