Fix XML namespace syntax (3.1.1 port of 3905)
The XML Object's namespace field was changed from "URL"
to "absolute URI" because relative references in a namespace
are deprecated by XML, and the base URI to use for resolving
them in the context of an OpenAPI Description is unclear.

However, XML namespaces can include fragments, and the
correct term is "non-relative URI" rather than "absolute URI"
which forbids fragments.

This change includes additional guidance on how XML usage and
the requirements of this specification do not quite align.