format: byte also defaults to octet-stream (3.0.4)
The description of defaults in the `contentType` fixed fields
table in 3.0.3 did not match the description in the Media Type Object's
"Considerations for multipart" section.

The "considerations for multipart" section appears to be more complete
and correct in both 3.0.3 and 3.1.1, so this adjusts that table
that replaced the fixed field defaults description to match the
"considerations for multipart" section.