"Object" def, "with" in titles, fix level
Define capital-O "Object" to make it clear why it is only occasionally
capitalized.

Use lowercase "with" in titles consistently (it was more common than
capitalized "With".  This is one of those rules that is different
depending on whose style guide you use- about half the major guides
say always lowercase, the other say always captial.  We were using
lowercase more often.

Also, fix "Working with Examples" to be a subsection of the
Example Object as otherwise it breaks the pattern of all of the
headings of its level being Object definitions.  I remember messing
with this a lot when I first posted that PR and either I just
missed this or I had some reason I have now forgotten to do it
this way.