Generalize registry site to publications
The spec site main page had gotten a bit muddled.  This organizes
it into three distinct sections (Specifications, Registries, and
Non-normative JSON Schemas) and adds an intro paragraph that links
to each section as well as the Learn and Tools site.

The reason for moving the schemas is that I almost couldn't find
the Overlay specification (I added Overlay to the shortcut links
at the top as well as it was missing).

The registries are in-between as they are less-well-known,
and I added shortcuts to the individual registries as it made
the sections feel more similar.

I renamed the overall site "Publications" instead of "Registry",
as it does a lot more now.  It could be "Specifications" if we
would rather stick with the "spec site" terminology.