Strawperson proposal for URI-based tag resolution
Tag resolution is difficult to do with URIs because tags are ordered
in the OpenAPI Object, and the order is significant.

This change takes the approach of adding an OpenAPI Document URI
to determine which OpenAPI Object is used to look up the named tag.

In the Operation Object, this is a parallel field called `tagRefs`,
that groups tags under URIs.  In the Tag Object, the `parentRef`
field provides the context for the existing `parent` field.