First pass at a media types registry.
This attempts to explain the basics of what is going on
in terms of "modeling" (what you do with a Schema Object
and in-memory data structures) and "mapping" (how you translate
that model into something that fits a particular media type
that can be serialized onto, or parsed off of, the wire).

This includes some sequential / streaming formats that will
be added for 3.2, as well as a selection of media types
directly addressed in the OAS text.  We could add more,
for example `image/*` that is more-or-less identical
to `application/octet-stream`, but this gives the general idea.

I omitted YAML since AFAIK it is rarely used in actual API payloads,
and such use is not directly mentioned in the OAS.