Media Types Registry
This registry groups media types by their data modeling
approaches, and provides pages that give a brief overview
of that approach as well as links to relevant OAS sections.
The section titles and anchors are intended to work with
3.2 (so we should not publish it until 3.2 is published).

The summary and remarks are only intended to give the general
idea of how the media types work, and what changes have
been made from release to release.

This initial set only includes media types that are mentioned
in the OAS, but the layout should work without needing any
links to such sections.  We'll have to decide how much detail
to put if and when we start making additions outside of the OAS,
which we hope to eventually do.  The point of this is to not
require a minor OAS version to add future media types, as long
as they do not require new fields or Objects.

I omitted YAML since AFAIK it is rarely used in actual API payloads,
and the OAS provides no guidance on how to handle YAML features that
are not compatible with JSON.  And of course for JSON-compatible
YAML, you just treat it like JSON.