Sequential media type support.
The fundamental approach in this change was proposed by Karen
Etheridge.

This adds "itemSchema" as a schema to apply to each entry in
a sequential media type instance.  It also defines how to map
sequential media tyes for use with "schema", and explains that
"schema" applies to complete content only.

JSON-based and text/event-stream media types are included.

Co-authored-by: Karen Etheridge <ether@cpan.org>