Add in: "querystring" parameter location
This allows treating the entire query string as a single parameter,
which for `application/x-www-form-urlencoded` results in it being
handled exactly as request bodies of that media type are handled.

Only one `in: "querystring"` is allowed per operation, and if it
is present, no `in: "query"` parameters are allowed.

The `content` field MUST be used for `in: "querystring"`, as the
`style` and related fields are handled in the Encoding Object.

This also includes a recommendation that implementors design
media types for complex query string formats such as those defined
by various frameworks, and register a procedure for supporting
them in our media type registry.  This is intended to address
a slow but steady trickle of requests to support various complex
and often contradictory query string formats.