Allow the Encoding Object in all locations
There is no reason to restrict use of the `encoding` field of
the Media Type Object to Request Body Objects only.  Not only
does this exclude legitimate use cases of returning multipart
content (particularly as we fix the support for `multipart/mixed`),
it also complicates implementation by requiring the handlingn of
the Media Type Object to be context-aware.

This is a rare change that both simplifies implementation and
adds functionality.