Define a policy for deprecation markings.
We already deprecate `example` in the Schema Object, but do not
define what deprecations means.  We also effectively deprecate
`allowEmptyValue` in the Parameter Object but did not explictly
say so.  We will likely make other deprecations in the future.

This defines a very basic deprecaton policy (they are still
supported, but we reserve the right to define a policy for
removing them in the future) and marks `allowEmptyValue` with
the same formatting used for `example`.