Provide guidance on null in XML.
There really isn't a native `null` type in XML, as both elements
and attributes that are empty have an empty string value.

We also need to leave the behavior implementation-defined
for compatibility.

However, the `xsi:nil` attribute is the closest thing to a `null`
element.  Attributes are harder, and the best I can come up with
is letting `null` behave the same as an omitted attribute for the
purpose of serialization.