The "?" is not part of individual parameter serializations
We say that the example table is "as would be shown with the
`example` or `examples` keywords", but then we include the `?`
which is not correct for a single parameter with `in: query`.

Even with `in: querystring`, it would not be correct because the
parameter describes the `application/x-www-form-urlencoded` document,
which does not include the `?` delimiter.